+++ docker/web/run.sh
@@ -7,6 +7,8 @@ mix event_store.init
 mix ecto.create
 mix ecto.setup
 
+MIX_ENV=test mix event_store.create
+MIX_ENV=test mix event_store.init
 mix test
 
 cd assets/ && yarn install && cd -