+++ lib/dealog_backoffice/app.ex
@@ -1,0 +1,6 @@
+
+defmodule DealogBackoffice.App do
+  use Commanded.Application, otp_app: :dealog_backoffice
+
+  router DealogBackoffice.Router
+end