+++ lib/dealog_backoffice/messages.ex
@@ -1,5 +1,0 @@
-defmodule DealogBackoffice.Messages do
-  @moduledoc """
-  The context for messages
-  """
-end